+++ start_remote.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env sh
+node lib/remote/remote.js