+++ changelog.md
@@ -4,7 +4,7 @@ Changelog
 1.1.0
 -----
 
-  * Added remote control
+  * Added remote control (remote.html)
   * Presenter is now configurable via drag & drop
   * Added notes
   * Added some print functionality